+++ example/so_flutter_example.dart
@@ -156,11 +156,12 @@ class _DataEntry extends DataScreen {
     // Create a double field.
     Field<double> numericField1 = doubleField(
       initialValue: 123.45,
+      decimals: 1,
       decoration: const InputDecoration(
-          labelText: 'Enter some decimal number that is less that 30000'),
+          labelText: 'Enter some decimal number that is less that 3000000'),
       validator: (v) {
-        return v == null || v >= 30000
-            ? 'Out of range! Should be less than 30000'
+        return v == null || v >= 3000000
+            ? 'Out of range! Should be less than 3000000'
             : null;
       },
       minValue: 25,
@@ -209,6 +210,10 @@ class _DataEntry extends DataScreen {
             numericField2,
             moneyField(
                 initialValue: Money(12345.23, 'USD'),
+                selectableCurrencies: [
+                  Currency.get('AED'),
+                  Currency.get('INR')
+                ],
                 decoration:
                     const InputDecoration(labelText: 'Amount')), // Money field
             comboField<String>(